remove delegation.go
